ccPairId-page-refactor: components breakdown, moved helpers and interfaces
