+++ web/src/components/adminPageComponents/single-connector/AdminSingleConnectorReIndexButton.tsx
@@ -4,7 +4,7 @@ import { PopupSpec, usePopup } from "@/components/adminPageComponents/connectors
 import { runConnector } from "@/lib/connector";
 import { Button, Divider, Text } from "@tremor/react";
 import { mutate } from "swr";
-import { buildCCPairInfoUrl } from "@/app/admin/connector/[ccPairId]/lib";
+import { buildCCPairInfoUrl } from "@/lib/connector/helpers";
 import { useState } from "react";
 import { Modal } from "@/components/Modal";
