+++ web/src/interfaces/connector/interfaces.ts
