+++ web/src/lib/connector/helpers.ts
